Prevent the browser from going back on backspace
This should be the default behavior to make text widgets usable.
At a later iteration it may be better to only disable the browser from
going back when a widget is selected that is expecting keyboard input
where backspace is intended for something else.

I expect this functionality would require some way of accessing the
web_sys event from within a druid callback.